Optimized and unified WASM crates
