+++ packages/hoprd/crates/hoprd-misc/Cargo.toml
@@ -13,11 +13,11 @@ crate-type = ["cdylib", "rlib"]
 
 [dependencies]
 real-base = { path = "../../../real/crates/real-base" }
-wasm-bindgen = { version = "0.2.81", features = [] }
+wasm-bindgen = "0.2.83"
 wasm-bindgen-test = "0.3.30"
 
-# Unfortunately, `wee_alloc` requires nightly Rust when targeting real for now.
-wee_alloc = { version = "0.4.5", optional = true }
+console_error_panic_hook = { version = "0.1.6", optional = true }
+#wee_alloc = { version = "0.4.5", optional = true }
 
 [dev-dependencies]
 wasm-bindgen-test = "0.3.30"
