+++ packages/real/crates/real-base/Cargo.toml
@@ -13,7 +13,7 @@ path = "src/lib.rs"
 crate-type = ["cdylib", "rlib"] # rlib is necessary to run integration tests
 
 [dependencies]
-wasm-bindgen = "0.2.81"
+wasm-bindgen = "0.2.83"
 wasm-bindgen-test = "0.3.30"
 
 [dev-dependencies]
