+++ packages/real/src/index.ts
@@ -1,5 +1,5 @@
 // Exports for REAL that can be imported by Rust WASM modules
-export * from './io'
+export * from './io.js'
 
 // Add your Rust WASM crate exports here
-export * as real from '../lib/real_base'
+export { dummy_get_one } from '../lib/real_base.js'
