+++ packages/utils/crates/utils-misc/Cargo.toml
@@ -15,11 +15,10 @@ crate-type = ["cdylib", "rlib"]
 real-base = { path = "../../../real/crates/real-base" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
-wasm-bindgen = { version = "0.2.81", features = ["serde-serialize"] }
-wasm-bindgen-test = "0.3.30"
+wasm-bindgen = { version = "0.2.83", features = ["serde-serialize"] }
 
-# Unfortunately, `wee_alloc` requires nightly Rust when targeting real for now.
-wee_alloc = { version = "0.4.5", optional = true }
+console_error_panic_hook = { version = "0.1.6", optional = true }
+#wee_alloc = { version = "0.4.5", optional = true }
 
 [dev-dependencies]
 wasm-bindgen-test = "0.3.30"
